third_party/WebKit/LayoutTests/fast/dom/HTMLAnchorElement/set-text.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/text-level-semantics.html#dom-a-text">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that HTMLAnchorElement.text can be set.");

var a = document.createElement("a");
shouldBe("a.__proto__", "HTMLAnchorElement.prototype");

shouldBeEqualToString("a.textContent", "");
shouldBeEqualToString("a.text", "");

a.textContent = "text1";
shouldBeEqualToString("a.textContent", "text1");
shouldBeEqualToString("a.text", "text1");

a.text = "text2";
shouldBeEqualToString("a.textContent", "text2");
shouldBeEqualToString("a.text", "text2");
</script>
</body>
</html>
