third_party/WebKit/LayoutTests/fast/dom/HTMLAnchorElement/set-text-expected.txt

Tests that HTMLAnchorElement.text can be set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS a.__proto__ is HTMLAnchorElement.prototype
PASS a.textContent is ""
PASS a.text is ""
PASS a.textContent is "text1"
PASS a.text is "text1"
PASS a.textContent is "text2"
PASS a.text is "text2"
PASS successfullyParsed is true

TEST COMPLETE
